Fix warning on Mac compilation

+++ upnp/src/threadutil/ThreadPool.c
@@ -428,7 +428,7 @@ static void SetSeed(void)
 		PtrToUint(ithread_get_current_thread_id().p));
 #elif defined(BSD) || defined(__APPLE__) || defined(__FreeBSD_kernel__)
 	srand((unsigned int)t.tv_usec +
-		(unsigned int)ithread_get_current_thread_id());
+		(unsigned int)(unsigned long)ithread_get_current_thread_id());
 #elif defined(__linux__) || defined(__sun) || defined(__CYGWIN__) || \
 	defined(__GLIBC__)
 	srand((unsigned int)t.tv_usec +
